requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
matplotlib==3.8.2
psycopg2-binary
